Removed DatePicker from defaults.json; moved copy.json from DateTimeInput to DatePicker; moved states from TimeInput to DatePicker; added states and moved most of the logic from DateTimeInput to DatePicker #907
