+++ src/defaults.json
@@ -35,226 +35,6 @@
     {
         "currency" : "USD"
     },
-	"DatePicker": {
-		"month": "February",
-		"year": "2018",
-		"label": "January 2000",
-		"headers": [{
-			"label": "Mon",
-			"title": "Monday"
-		}, {
-			"label": "Tue",
-			"title": "Tuesday"
-		}, {
-			"label": "Wed",
-			"title": "Wednesday"
-		}, {
-			"label": "Thu",
-			"title": "Thursday"
-		}, {
-			"label": "Fri",
-			"title": "Friday"
-		}, {
-			"label": "Sat",
-			"title": "Saturday"
-		}, {
-			"label": "Sun",
-			"title": "Sunday"
-		}],
-		"items": [
-			[{
-				"label": "01",
-				"value": "1",
-				"isCurrent": false,
-				"isDisabled": true,
-				"isSelected": false
-			}, {
-				"label": "02",
-				"value": "2",
-				"isCurrent": false,
-				"isDisabled": true,
-				"isSelected": false
-			}, {
-				"label": "03",
-				"value": "3",
-				"isCurrent": false,
-				"isDisabled": true,
-				"isSelected": false
-			}, {
-				"label": "04",
-				"value": "4",
-				"isCurrent": false,
-				"isDisabled": true,
-				"isSelected": false
-			}, {
-				"label": "05",
-				"value": "5",
-				"isCurrent": false,
-				"isDisabled": true,
-				"isSelected": false
-			}, {
-				"label": "06",
-				"value": "6",
-				"isCurrent": false,
-				"isDisabled": true,
-				"isSelected": false
-			}, {
-				"label": "07",
-				"value": "7",
-				"isCurrent": false,
-				"isDisabled": true,
-				"isSelected": false
-			}],
-			[{
-				"label": "08",
-				"value": "8",
-				"isCurrent": false,
-				"isDisabled": true,
-				"isSelected": false
-			}, {
-				"label": "09",
-				"value": "9",
-				"isCurrent": false,
-				"isDisabled": true,
-				"isSelected": false
-			}, {
-				"label": "10",
-				"value": "10",
-				"isCurrent": false,
-				"isDisabled": true,
-				"isSelected": false
-			}, {
-				"label": "11",
-				"value": "11",
-				"isCurrent": false,
-				"isDisabled": true,
-				"isSelected": false
-			}, {
-				"label": "12",
-				"value": "12",
-				"isCurrent": false,
-				"isDisabled": true,
-				"isSelected": false
-			}, {
-				"label": "13",
-				"value": "13",
-				"isCurrent": false,
-				"isDisabled": true,
-				"isSelected": false
-			}, {
-				"label": "14",
-				"value": "14",
-				"isCurrent": false,
-				"isDisabled": true,
-				"isSelected": false
-			}],
-			[{
-				"label": "15",
-				"value": "15",
-				"isCurrent": false,
-				"isDisabled": true,
-				"isSelected": false
-			}, {
-				"label": "16",
-				"value": "16",
-				"isCurrent": false,
-				"isDisabled": true,
-				"isSelected": false
-			}, {
-				"label": "17",
-				"value": "17",
-				"isCurrent": false,
-				"isDisabled": true,
-				"isSelected": false
-			}, {
-				"label": "18",
-				"value": "18",
-				"isCurrent": false,
-				"isDisabled": true,
-				"isSelected": false
-			}, {
-				"label": "19",
-				"value": "19",
-				"isCurrent": false,
-				"isDisabled": true,
-				"isSelected": false
-			}, {
-				"label": "20",
-				"value": "20",
-				"isCurrent": false,
-				"isDisabled": true,
-				"isSelected": false
-			}, {
-				"label": "21",
-				"value": "21",
-				"isCurrent": false,
-				"isDisabled": true,
-				"isSelected": false
-			}],
-			[{
-				"label": "22",
-				"value": "22",
-				"isCurrent": false,
-				"isDisabled": true,
-				"isSelected": false
-			}, {
-				"label": "23",
-				"value": "23",
-				"isCurrent": false,
-				"isDisabled": true,
-				"isSelected": false
-			}, {
-				"label": "24",
-				"value": "24",
-				"isCurrent": true,
-				"isDisabled": false,
-				"isSelected": true
-			}, {
-				"label": "25",
-				"value": "25",
-				"isCurrent": false,
-				"isDisabled": false,
-				"isSelected": false
-			}, {
-				"label": "26",
-				"value": "26",
-				"isCurrent": false,
-				"isDisabled": false,
-				"isSelected": false
-			}, {
-				"label": "27",
-				"value": "27",
-				"isCurrent": false,
-				"isDisabled": false,
-				"isSelected": false
-			}, {
-				"label": "28",
-				"value": "28",
-				"isCurrent": false,
-				"isDisabled": false,
-				"isSelected": false
-			}],
-			[{
-				"label": "29",
-				"value": "29",
-				"isCurrent": false,
-				"isDisabled": false,
-				"isSelected": false
-			}, {
-				"label": "30",
-				"value": "30",
-				"isCurrent": false,
-				"isDisabled": false,
-				"isSelected": false
-			}, {
-				"label": "31",
-				"value": "31",
-				"isCurrent": false,
-				"isDisabled": false,
-				"isSelected": false
-			}]
-		]
-	},
 	"Grid": {
 		"children": "{<GridItem><Text>First column</Text></GridItem><GridItem><Text>Second column</Text></GridItem><GridItem><Text>Third column</Text></GridItem>}"
 	},
